+++ src/launchpad/cli.py
@@ -1,30 +1,16 @@
 from __future__ import annotations
 
 import asyncio
-import json
 import logging
 import os
-import time
-from pathlib import Path
-from typing import cast
 
 import click
-from rich.console import Console
-from rich.progress import Progress, SpinnerColumn, TextColumn
-from rich.table import Table
 
 from . import __version__
-from .analyzers.android import AndroidAnalyzer
-from .analyzers.apple import AppleAppAnalyzer
-from .artifacts.artifact import AndroidArtifact, AppleArtifact
-from .artifacts.artifact_factory import ArtifactFactory
-from .models.android import AndroidAnalysisResults
-from .models.apple import AppleAnalysisResults
-from .models.common import FileAnalysis
 from .service import run_service
+from .size.cli import size_command
+from .utils.console import console
 from .utils.logging import setup_logging
-
-console = Console()
 
 
 @click.group(invoke_without_command=True)
@@ -38,191 +24,6 @@ def cli(ctx: click.Context, version: bool) -> None:
 
     if ctx.invoked_subcommand is None:
         click.echo(ctx.get_help())
-
-
-@cli.command()
-@click.argument("input_path", type=click.Path(exists=True, path_type=Path), metavar="INPUT_PATH")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    default="apple-app-analysis-report.json",
-    help="Output path for the JSON analysis report.",
-    show_default=True,
-)
-@click.option(
-    "--working-dir",
-    type=click.Path(path_type=Path),
-    help="Working directory for temporary files (default: system temp).",
-)
-@click.option("--skip-swift-metadata", is_flag=True, help="Skip Swift metadata parsing for faster analysis.")
-@click.option("--skip-symbols", is_flag=True, help="Skip symbol extraction and analysis.")
-@click.option("--skip-range-mapping", is_flag=True, help="Skip range mapping for binary content categorization.")
-@click.option("--skip-treemap", is_flag=True, help="Skip treemap generation for hierarchical size analysis.")
-@click.option("--verbose", "-v", is_flag=True, help="Enable verbose logging output.")
-@click.option("--quiet", "-q", is_flag=True, help="Suppress all output except errors.")
-@click.option(
-    "--format",
-    "output_format",
-    type=click.Choice(["json", "table"], case_sensitive=False),
-    default="json",
-    help="Output format for results.",
-    show_default=True,
-)
-def apple_app(
-    input_path: Path,
-    output: Path,
-    working_dir: Path | None,
-    skip_swift_metadata: bool,
-    skip_symbols: bool,
-    skip_range_mapping: bool,
-    skip_treemap: bool,
-    verbose: bool,
-    quiet: bool,
-    output_format: str,
-) -> None:
-    """Analyze an Apple app bundle and generate a size report.
-
-    INPUT_PATH can be:
-    - .xcarchive.zip file
-    """
-
-    setup_logging(verbose=verbose, quiet=quiet)
-
-    if verbose and quiet:
-        raise click.UsageError("Cannot specify both --verbose and --quiet")
-
-    _validate_apple_input(input_path)
-
-    if not quiet:
-        console.print(f"[bold blue]App Size Analyzer v{__version__}[/bold blue]")
-        console.print(f"Analyzing Apple app: [cyan]{input_path}[/cyan]")
-        console.print(f"Output: [cyan]{output}[/cyan]")
-        console.print()
-
-    try:
-        start_time = time.time()
-
-        with Progress(
-            SpinnerColumn(),
-            TextColumn("[progress.description]{task.description}"),
-            console=console,
-            disable=quiet,
-        ) as progress:
-            task = progress.add_task("Analyzing Apple app bundle...", total=None)
-
-            analyzer = AppleAppAnalyzer(
-                working_dir=working_dir,
-                skip_swift_metadata=skip_swift_metadata,
-                skip_symbols=skip_symbols,
-                skip_range_mapping=skip_range_mapping,
-                skip_treemap=skip_treemap,
-            )
-            artifact = ArtifactFactory.from_path(input_path)
-            analyzer.preprocess(cast(AppleArtifact, artifact))
-            results = analyzer.analyze(cast(AppleArtifact, artifact))
-
-            progress.update(task, description="Analysis complete!")
-
-        end_time = time.time()
-        duration = end_time - start_time
-
-        results = results.model_copy(update={"analysis_duration": duration})
-
-        if output_format == "json":
-            _write_json_output(results, output, quiet)
-        else:
-            _print_apple_table_output(results, quiet)
-
-        if not quiet:
-            console.print(f"\n[bold green]✓[/bold green] Analysis completed in {duration:.2f}s")
-            _print_apple_summary(results)
-
-    except Exception as e:
-        if verbose:
-            console.print_exception()
-        else:
-            console.print(f"[bold red]Error:[/bold red] {e}")
-        raise click.Abort()
-
-
-@cli.command()
-@click.argument("input_path", type=click.Path(exists=True, path_type=Path), metavar="INPUT_PATH")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    default="android-analysis-report.json",
-    help="Output path for the JSON analysis report.",
-    show_default=True,
-)
-@click.option("--verbose", "-v", is_flag=True, help="Enable verbose logging output.")
-@click.option("--quiet", "-q", is_flag=True, help="Suppress all output except errors.")
-@click.option(
-    "--format",
-    "output_format",
-    type=click.Choice(["json", "table"], case_sensitive=False),
-    default="json",
-    help="Output format for results.",
-    show_default=True,
-)
-def android(
-    input_path: Path,
-    output: Path,
-    verbose: bool,
-    quiet: bool,
-    output_format: str,
-) -> None:
-    """Analyze an Android app bundle and generate a size report.
-
-    INPUT_PATH can be:
-    - Android .apk file
-    - Android .aab file (coming soon)
-    """
-    setup_logging(verbose=verbose, quiet=quiet)
-
-    if verbose and quiet:
-        raise click.UsageError("Cannot specify both --verbose and --quiet")
-
-    if not quiet:
-        console.print(f"[bold blue]App Size Analyzer v{__version__}[/bold blue]")
-        console.print(f"Analyzing Android app: [cyan]{input_path}[/cyan]")
-        console.print(f"Output: [cyan]{output}[/cyan]")
-        console.print()
-
-    try:
-        start_time = time.time()
-
-        with Progress(
-            SpinnerColumn(),
-            TextColumn("[progress.description]{task.description}"),
-            console=console,
-            disable=quiet,
-        ) as progress:
-            task = progress.add_task("Analyzing Android app bundle...", total=None)
-
-            analyzer = AndroidAnalyzer()
-            artifact = ArtifactFactory.from_path(input_path)
-            results = analyzer.analyze(cast(AndroidArtifact, artifact))
-
-            progress.update(task, description="Analysis complete!")
-
-        end_time = time.time()
-        duration = end_time - start_time
-
-        results = results.model_copy(update={"analysis_duration": duration})
-
-        if output_format == "json":
-            _write_json_output(results, output, quiet)
-        else:
-            _print_android_table_output(results, quiet)
-
-    except Exception as e:
-        if verbose:
-            console.print_exception()
-        else:
-            console.print(f"[bold red]Error:[/bold red] {e}")
-        raise click.Abort()
 
 
 @cli.command()
@@ -279,137 +80,7 @@ def serve(host: str, port: int, mode: str | None, verbose: bool) -> None:
         raise click.Abort()
 
 
-def _validate_apple_input(input_path: Path) -> None:
-    """Validate that the input path looks like an Apple artifact."""
-    suffix = input_path.suffix.lower()
-    valid_extensions = {".zip"}
-
-    if suffix not in valid_extensions:
-        raise click.BadParameter(
-            f"'{input_path}' doesn't look like a typical Apple artifact. "
-            f"Expected one of: {', '.join(sorted(valid_extensions))}"
-        )
-
-
-def _write_json_output(results: AppleAnalysisResults | AndroidAnalysisResults, output_path: Path, quiet: bool) -> None:
-    """Write results to JSON file."""
-    # Ensure output directory exists
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-
-    # Write JSON with proper formatting
-    with open(output_path, "w", encoding="utf-8") as f:
-        json.dump(results.to_dict(), f, indent=2, ensure_ascii=False)
-
-    if not quiet:
-        console.print(f"[bold green]✓[/bold green] Results written to: [cyan]{output_path}[/cyan]")
-
-
-def _print_apple_table_output(results: AppleAnalysisResults, quiet: bool) -> None:
-    """Print results in table format to console."""
-    if quiet:
-        return
-
-    # App Info Table
-    app_table = Table(title="App Information", show_header=True, header_style="bold magenta")
-    app_table.add_column("Property", style="cyan")
-    app_table.add_column("Value", style="white")
-
-    app_info = results.app_info
-    app_table.add_row("Name", app_info.name)
-    app_table.add_row("Bundle ID", app_info.bundle_id)
-    app_table.add_row("Version", f"{app_info.version} ({app_info.build})")
-    app_table.add_row("Min OS", app_info.minimum_os_version)
-    app_table.add_row("Platforms", ", ".join(app_info.supported_platforms))
-
-    console.print(app_table)
-    console.print()
-
-    file_analysis = results.file_analysis
-    _print_file_analysis_table(file_analysis)
-
-    # File Types Table
-    if file_analysis.file_type_sizes:
-        type_table = Table(title="File Types", show_header=True, header_style="bold yellow")
-        type_table.add_column("Type", style="cyan")
-        type_table.add_column("Size", style="white")
-        type_table.add_column("Percentage", style="green")
-
-        total_size = file_analysis.total_size
-        for file_type, size in sorted(file_analysis.file_type_sizes.items(), key=lambda x: x[1], reverse=True)[
-            :10
-        ]:  # Top 10 file types
-            percentage = (size / total_size) * 100 if total_size > 0 else 0
-            type_table.add_row(file_type or "unknown", _format_bytes(size), f"{percentage:.1f}%")
-
-        console.print(type_table)
-
-
-def _print_android_table_output(results: AndroidAnalysisResults, quiet: bool) -> None:
-    """Print results in table format to console."""
-    if quiet:
-        return
-
-    # App Info Table
-    app_table = Table(title="App Information", show_header=True, header_style="bold magenta")
-    app_table.add_column("Property", style="cyan")
-    app_table.add_column("Value")
-
-    app_info = results.app_info
-    app_table.add_row("Name", app_info.name)
-    app_table.add_row("Package Name", app_info.package_name)
-    app_table.add_row("Version", f"{app_info.version} ({app_info.build})")
-
-    console.print(app_table)
-    console.print()
-
-    _print_file_analysis_table(results.file_analysis)
-
-
-def _print_file_analysis_table(file_analysis: FileAnalysis) -> None:
-    file_table = Table(title="File Analysis", show_header=True, header_style="bold green")
-    file_table.add_column("Metric", style="cyan")
-    file_table.add_column("Value")
-
-    file_table.add_row("Total Size", _format_bytes(file_analysis.total_size))
-    file_table.add_row("File Count", str(file_analysis.file_count))
-
-    console.print(file_table)
-    console.print()
-
-
-def _print_apple_summary(results: AppleAnalysisResults) -> None:
-    """Print a brief summary of the analysis."""
-    file_analysis = results.file_analysis
-    binary_analysis = results.binary_analysis
-    insights = results.insights
-
-    console.print("\n[bold]Summary:[/bold]")
-    console.print(f"• App name: [cyan]{results.app_info.name}[/cyan]")
-    console.print(f"• Total app size: [cyan]{_format_bytes(file_analysis.total_size)}[/cyan]")
-    console.print(f"• File count: [cyan]{file_analysis.file_count:,}[/cyan]")
-
-    if insights and insights.duplicate_files and insights.duplicate_files.total_savings > 0:
-        console.print(
-            f"• Potential savings from duplicates: "
-            f"[yellow]{_format_bytes(insights.duplicate_files.total_savings)}[/yellow]"
-        )
-
-    if binary_analysis:
-        for binary in binary_analysis:
-            console.print(f"\nExecutable Size: {binary.executable_size / 1024 / 1024:.1f} MB")
-            console.print(f"Architectures: {', '.join(binary.architectures)}")
-            console.print(f"Linked Libraries: {len(binary.linked_libraries)}")
-            console.print(f"Sections: {len(binary.sections)}")
-
-
-def _format_bytes(size: int) -> str:
-    """Format byte size in human-readable format."""
-    size_float = float(size)
-    for unit in ["B", "KB", "MB", "GB"]:
-        if size_float < 1024.0:
-            return f"{size_float:.1f} {unit}"
-        size_float /= 1024.0
-    return f"{size_float:.1f} TB"
+cli.add_command(size_command)
 
 
 def main() -> None:
